pyproject.toml
[project]
name = "webjango"
version = "0.1.0"
description = "Django app Raz"
authors = [
    {name = "Raz"}
]
license = {text = "Free"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "django-stubs[compatible-mypy] (>=5.2.8,<6.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]

[tool.ruff]
line-length = 120

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
DJANGO_SETTINGS_MODULE = "raz_webjango.settings"

[pytest]
DJANGO_SETTINGS_MODULE = "raz_webjango.settings"